requirements.txt
Flask 
keras 
matplotlib
scikit-image
tensorflow
torch
pillow
numpy
torchvision